Tenancy: Implement __FALLBACK tenant and __FALLBACK privilege (#788)
* Implement __FALLBACK tenant and __FALLBACK privileges.
* Add unit test to test vmdkops on VM which belongs to __FALLBACK tenant.
* Modify function "create_vm" and "remove_vm" to avoid calling function "find_vm_by_name".
   Remove function "find_vm_by_name" since vm name is not unique and we cannot find vm
by name.
* Fix a bug in function "convert" in kvESX.py.
There is a python2 and python 3 compatible issue with division.

+++ esx_service/utils/kvESX.py
@@ -360,8 +360,8 @@ def convert(size):
     if size < KB:
         return size
     elif size < MB:
-        return '{0}{1}'.format(size / KB, 'KB')
+        return '{0}{1}'.format(int(size / KB), 'KB')
     elif size < GB:
-        return '{0}{1}'.format(size / MB, 'MB')
+        return '{0}{1}'.format(int(size / MB), 'MB')
     else:
-        return '{0}{1}'.format(size / GB, 'GB')
+        return '{0}{1}'.format(int(size / GB), 'GB')
